some warnings are fixed

+++ FBXModule/FBXModule_IO.cpp
@@ -31,7 +31,7 @@ public:
         :
         orgDir(GetCurrentDirectory(0, nullptr), 0)
     {
-        GetCurrentDirectory(orgDir.length(), orgDir.data());
+        GetCurrentDirectory(DWORD(orgDir.length()), orgDir.data());
         SetCurrentDirectory(changeDir.c_str());
     }
     ~_DirectoryModifier(){
